+++ src/Facet/Elab/Term.hs
@@ -79,6 +79,7 @@ import           Facet.Snoc
 import           Facet.Source (Source)
 import qualified Facet.Surface as S
 import           Facet.Syntax
+import           Facet.Unify
 import           Facet.Usage hiding (restrict)
 import           GHC.Stack
